requirements.txt
scikit-learn
pandas
numpy
setuptools
PyYAML
streamlit
-e .